Fix Prevent Dynamic Code Execution
In file: module_ops.py, untrusted, user-provided data is executed dynamically. This allows an attacker to execute arbitrary code or at least modify the code that should be executed. iCR identified the code where the dynamic code execution happens and suggested that the user-provided data should be sanitized before using it as code.